+++ OktaPosh/public/User.ps1
@@ -1,3 +1,56 @@
+# https://developer.okta.com/docs/reference/api/users/
+
+function New-OktaAuthProviderUser
+{
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSShouldProcess", "")]
+    [CmdletBinding(SupportsShouldProcess)]
+    param(
+        [Parameter(Mandatory,ValueFromPipelineByPropertyName)]
+        [Alias("given_name")]
+        [string] $FirstName,
+        [Parameter(Mandatory,ValueFromPipelineByPropertyName)]
+        [Alias("family_name")]
+        [string] $LastName,
+        [Parameter(Mandatory,ValueFromPipelineByPropertyName)]
+        [string] $Email,
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [string] $Login,
+        [Parameter(Mandatory)]
+        [ValidateSet('OKTA', 'ACTIVE_DIRECTORY', 'LDAP', 'FEDERATION', 'SOCIAL', 'IMPORT')]
+        [string] $ProviderType,
+        [string] $ProviderName,
+        [string[]] $GroupIds
+    )
+
+    process {
+        Set-StrictMode -Version Latest
+
+        if (!$Login) {
+            $Login = $Email
+        }
+        $body = @{
+            profile = @{
+                firstName = $FirstName
+                lastName = $LastName
+                email = $Email
+                login = $Login
+              }
+            credentials = @{
+                provider = @{
+                  type = $ProviderType
+                }
+              }
+        }
+        if ($ProviderName) {
+            $body.credentials.provider['name'] = $ProviderName
+        }
+        if ($GroupIds) {
+            $body['groupIds'] = @($GroupIds)
+        }
+        Invoke-OktaApi -RelativeUri "users?provider=true" -Body $body -Method POST
+    }
+}
+
 function New-OktaUser
 {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSShouldProcess", "")]
@@ -52,9 +105,8 @@ function Get-OktaUser {
         [Parameter(ParameterSetName="Query")]
         [Parameter(ParameterSetName="Search")]
         [uint32] $Limit,
-        [Parameter(ParameterSetName="Query")]
-        [Parameter(ParameterSetName="Search")]
-        [string] $After,
+        [Parameter(ParameterSetName="Next")]
+        [switch] $Next,
         [Parameter(ParameterSetName="Search")]
         [string] $Search,
         [Parameter(ParameterSetName="Search")]
@@ -70,12 +122,47 @@ function Get-OktaUser {
             Invoke-OktaApi -RelativeUri "users/$UserId" -Json:$Json
         } else {
             Invoke-OktaApi -RelativeUri "users$(Get-QueryParameters `
-                                 -Query $Query -Limit $Limit -After $After `
-                                 -Filter $Filter `
-                                 -Search $Search -SortBy $SortBy -SortOrder $SortOrder `
-                                 )" -Json:$Json
+                                -Query $Query -Limit $Limit `
+                                -Filter $Filter `
+                                -Search $Search -SortBy $SortBy -SortOrder $SortOrder `
+                                )" -Json:$Json -Next:$Next
         }
     }
+}
+
+function Get-OktaUserApplication {
+    [CmdletBinding(DefaultParameterSetName="Limit")]
+    param (
+        [Parameter(Mandatory,ValueFromPipeline,ValueFromPipelineByPropertyName)]
+        [Alias("Id")]
+        [Alias("Login")]
+        [string] $UserId,
+        [Parameter(ParameterSetName="Limit")]
+        [uint32] $Limit,
+        [Parameter(ParameterSetName="Next")]
+        [switch] $Next,
+        [switch] $Json
+    )
+
+    $query = Get-QueryParameters -Filter "user.id eq `"$UserId`"" -Limit $Limit
+    Invoke-OktaApi -RelativeUri "apps$query&expand=user%2F$UserId" -Json:$Json -Next:$Next
+}
+
+function Get-OktaUserGroup {
+    [CmdletBinding(DefaultParameterSetName="Limit")]
+    param (
+        [Parameter(Mandatory,ValueFromPipeline,ValueFromPipelineByPropertyName)]
+        [Alias("Id")]
+        [Alias("Login")]
+        [string] $UserId,
+        [Parameter(ParameterSetName="Limit")]
+        [uint32] $Limit,
+        [Parameter(ParameterSetName="Next")]
+        [switch] $Next,
+        [switch] $Json
+    )
+
+    Invoke-OktaApi -RelativeUri "users/$UserId/groups$(Get-QueryParameters -Limit $Limit)" -Json:$Json -Next:$Next
 }
 
 function Remove-OktaUser {
@@ -102,6 +189,3 @@ function Remove-OktaUser {
 
     }
 }
-
-
-