+++ tests/New-RBR-UI-Configuration.ps1
@@ -1,8 +1,7 @@
 [CmdletBinding(SupportsShouldProcess)]
 param(
     [ValidateSet("dev","qa","uat","ct","int","prod")]
-    [string] $Environment = "dev",
-    [switch] $DataCapture
+    [string] $Environment
 )
 Set-StrictMode -Version Latest
 $ErrorActionPreference = "Stop"
@@ -17,10 +16,6 @@ if ($Environment -ne 'prod') {
 
 $name = "RBR"
 $uiPath = "fp-ui"
-if ($DataCapture) {
-    $name = "DataCapture"
-    $uiPath = "dc-ui"
-}
 $nameLower = $name.ToLowerInvariant()
 
 # script to add Okta object for the Reliance project
@@ -29,137 +24,92 @@ if (!(Get-Module OktaPosh)) {
     return
 }
 
-# AuthorizationServer configuration
-$authServerName = "Casualty-$name-AS"
-$scopes = @("casualty.$nameLower.client.usaa",
-            "casualty.$nameLower.client.den1",
-            "casualty.$nameLower.client.safeco",
-            "casualty.$nameLower.client.ins1", # this is junk test one
-            "casualty.$nameLower.client.nw")
-
-$claims = @(
-    @{
-        name = "roles"
-        valueType = "GROUPS"
-        groupFilterType = "STARTS_WITH"
-        value = "CCC-$name-Role-"
-        claimType= "ACCESS_TOKEN"
-    },
-    @{
-        name = "clients"
-        valueType = "GROUPS"
-        groupFilterType = "STARTS_WITH"
-        value = "CCC-$name-Client-"
-        claimType= "ACCESS_TOKEN"
-    },
-    @{
-        name = "email"
-        valueType = "EXPRESSION"
-        value = "user.email"
-        claimType= "ACCESS_TOKEN"
-    },
-    @{
-        name = "friendlyName"
-        valueType = "EXPRESSION"
-        value = 'String.len(user.displayName) > 0 ? user.displayName : user.firstName+ " " + user.lastName'
-        claimType= "ACCESS_TOKEN"
-    },
-    @{
-        name = "friendlyName"
-        valueType = "EXPRESSION"
-        value = 'String.len(user.displayName) > 0 ? user.displayName : user.firstName+ " " + user.lastName'
-        claimType= "ID_TOKEN"
-    },
-    @{
-        name = "login"
-        valueType = "EXPRESSION"
-        value = "user.email"
-        claimType= "ACCESS_TOKEN"
-    },
-    @{
-        name = "pictureUrl"
-        valueType = "EXPRESSION"
-        value = "appuser.picture"
-        claimType= "ID_TOKEN"
-    },
-    @{
-        name = "profileUrl"
-        valueType = "EXPRESSION"
-        value = "user.profileUrl"
-        claimType= "ID_TOKEN"
-    }
-)
-
-# Application Configuration
-$audience = "https://reliance/$uiPath"
-$description = "$name UI for casualty"
-$apps = @(
-    @{ Name = "CCC-CAS$name-SPA"
-       RedirectUris = @(
-        "https://reliance$domainSuffix.reprice.nhr.com/$uiPath/implicit/callback",
-        "http://localhost:8080/$uiPath/implicit/callback"
-        )
-       LoginUri = "https://reliance$domainSuffix.reprice.nhr.com/$uiPath/"
-       PostLogoutUris = "https://reliance$domainSuffix.reprice.nhr.com/$uiPath/"
-       Scopes = $scopes + "openid","profile","email"
-    }
-)
-
-# Groups
-$groupNames = @("CCC-$name-Client-USAA-Group",
-                "CCC-$name-Client-NW-Group",
-                "CCC-$name-Client-INS1-Group",
-                "CCC-$name-Client-GEICO-Group")
-
-# Origin
-$origins = @("https://reliance$domainSuffix.reprice.nhr.com")
-
-try {
-
-    . (Join-Path $PSScriptRoot New-OktaAuthServerConfig.ps1)
-    . (Join-Path $PSScriptRoot New-OktaAppConfig.ps1)
-
-    $authServer = New-OktaAuthServerConfig -authServerName $authServerName `
-                            -Scopes $scopes `
-                            -audience $audience `
-                            -description $description `
-                            -claims $claims
-
-    $groups = @()
-    foreach ($group in $groupNames) {
-        $g = Get-OktaGroup -Query $group
-        if ($g) {
-            Write-Host "Found group '$group'"
-        } else {
-            $g = New-OktaGroup -Name $group
-            Write-Host "Added group '$group'"
+$AuthServer = @{
+    authServerName = "Casualty-$name-AS"
+    audience = "https://reliance/$uiPath"
+    description = "$name UI for casualty"
+    scopes = @("casualty.$nameLower.client.usaa",
+                "casualty.$nameLower.client.den1",
+                "casualty.$nameLower.client.safeco",
+                "casualty.$nameLower.client.ins1", # this is junk test one
+                "casualty.$nameLower.client.nw")
+            
+    claims = @(
+        @{
+            name = "roles"
+            valueType = "GROUPS"
+            groupFilterType = "STARTS_WITH"
+            value = "CCC-$name-Role-"
+            claimType= "ACCESS_TOKEN"
+        },
+        @{
+            name = "clients"
+            valueType = "GROUPS"
+            groupFilterType = "STARTS_WITH"
+            value = "CCC-$name-Client-"
+            claimType= "ACCESS_TOKEN"
+        },
+        @{
+            name = "email"
+            valueType = "EXPRESSION"
+            value = "user.email"
+            claimType= "ACCESS_TOKEN"
+        },
+        @{
+            name = "friendlyName"
+            valueType = "EXPRESSION"
+            value = 'String.len(user.displayName) > 0 ? user.displayName : user.firstName+ " " + user.lastName'
+            claimType= "ACCESS_TOKEN"
+        },
+        @{
+            name = "friendlyName"
+            valueType = "EXPRESSION"
+            value = 'String.len(user.displayName) > 0 ? user.displayName : user.firstName+ " " + user.lastName'
+            claimType= "ID_TOKEN"
+        },
+        @{
+            name = "login"
+            valueType = "EXPRESSION"
+            value = "user.email"
+            claimType= "ACCESS_TOKEN"
+        },
+        @{
+            name = "pictureUrl"
+            valueType = "EXPRESSION"
+            value = "appuser.picture"
+            claimType= "ID_TOKEN"
+        },
+        @{
+            name = "profileUrl"
+            valueType = "EXPRESSION"
+            value = "user.profileUrl"
+            claimType= "ID_TOKEN"
         }
-        $groups += $g
-    }
-
-    foreach ($newApp in $apps) {
-        $app = New-OktaAppConfig -Name $newApp.Name `
-                        -Scopes $newApp.Scopes `
-                        -RedirectUris $newApp.RedirectUris `
-                        -LoginUri $newApp.LoginUri `
-                        -PostLogoutUris $newApp.PostLogoutUris `
-                        -GrantTypes "authorization_code", "password", "client_credentials", "implicit" `
-                        -AuthServerId $authServer.Id
-        foreach ($group in $groups) {
-            $null = Add-OktaApplicationGroup -AppId $app.id -GroupId $group.id
-            Write-Host "    Added '$($group.profile.name)' group to app '$($app.label)'"
-        }
-    }
-
-    foreach ($origin in $origins) {
-        if (Get-OktaTrustedOrigin -Filter "origin eq `"$origin`"") {
-            Write-Host "Found origin '$origin'"
-        } else {
-            $null = New-OktaTrustedOrigin -Name $origin -Origin $origin -CORS -Redirect
-            Write-Host "Added origin '$origin'"
-        }
-     }
-
-} catch {
-    Write-Error "$_`n$($_.ScriptStackTrace)"
+    )
 }
+
+$Applications = @(
+        @{ Name = "CCC-CAS$name-SPA"
+            RedirectUris = @(
+            "https://reliance$domainSuffix.reprice.nhr.com/$uiPath/implicit/callback",
+            "http://localhost:8080/$uiPath/implicit/callback"
+            )
+            LoginUri = "https://reliance$domainSuffix.reprice.nhr.com/$uiPath/"
+            PostLogoutUris = "https://reliance$domainSuffix.reprice.nhr.com/$uiPath/"
+            Scopes = '*'
+        }
+    )
+
+$GroupNames = @("CCC-$name-Client-DEN1-Group",
+                 "CCC-$name-Client-NW-Group",
+                 "CCC-$name-Client-SAFECO-Group",
+                 "CCC-$name-Client-USAA-Group",
+                 "CCC-$name-Client-INS1-Group" # test client in Reliance
+                 )
+
+$Origins = @("https://reliance$domainSuffix.reprice.nhr.com")
+
+. (Join-Path $PSScriptRoot New-OktaAppAndAuthServerConfig.ps1)
+
+New-OktaAppAndAuthServerConfig -AuthServer $AuthServer -Applications $Applications -GroupNames $GroupNames -Origins $Origins
+